requirements.txt
azure-core==1.5.0
azure-identity==1.3.1
azure-keyvault-secrets==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.9
isodate==0.6.0
msal==1.2.0
msal-extensions==0.1.3
msrest==0.6.13
oauthlib==3.1.0
portalocker==1.7.0
pycparser==2.20
PyJWT==1.7.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.9
